Demographic Analysis on the Evolution of Voters’ Participation and Preference
This project studies how the demongraphic profiles in each voting district influence the voters’
willingness to vote and their preferences. Instead of studying a single election event, this project
focuses on the shifts from 2012 election to 2016 election. The dataset includes a selected set of
counties in united states and their demographic information.